Changes
The ISBN was cutting off the last digit the user typed. I fixed my mistake. I also realized that the ISBN cuts off leading zeros when it converts to (long long). I wonder if there is a way to too the output stream to print leading zeroes, maybe a string format is better. :/